+++ src/utils/swiperSliders.ts
@@ -70,7 +70,7 @@ export const swiperSliders = () => {
         centeredSlides: true,
         grabCursor: true,
         autoplay: {
-          delay: 3000,
+          delay: 5000,
           disableOnInteraction: false,
         },
         navigation: {